- Removed convertUnixToTime() since it's dead code - moved formatTime() to dateTimeFormat.js and changed it to use Intl.DateTimeFormat() - moved formatSecondsFromMidnight() to /lib/dateTimeFormat.js and changed it to use Intl.DateTimeFormat() - moved functions from /lib/otp/dateTimeFormatters.js to /lib/dateTimeFormat.js - moved formatDepartureDisplay() from src/stores/tripOptionsStore.js to /lib/dateTimeFormat.js - moved formatTime() from src/components/ArrivalDeparture.svelte to /lib/dateTimeFormat.js and renamed it to formatTimeForArrivalDeparture() - moved Intl format from src/components/ArrivalDeparture.svelte to /lib/dateTimeFormat.js and renamed it to fourDigit24HourTimeFormat - added Temporal polyfill package - converted usages of Date() to Temporal equivalents - moved getTodayDate() and getCurrentTime() functions into /lib/dateTimeFormat.js and renamed to getTodayDateForInput() and getCurrentTimeForInput(). Also using Temporal instead of Date() - moved convertToISO8601() from graphql.js to /lib/dateTimeFormat.js - renamed format24HourTo12Hour to convert24HourTo12Hour - PR feedback addressed - ran prepush script to fix lint issues - second round of PR feedback: Fixed JSDoc descriptions, params, and examples. Added edge cases for convertToISO8601 unit tests - third round of PR feedback: Updated msToTimeString to default to local timezone instead of UTC. Created a getLocalTimeZone() function. Removed msToLocalTimeString() function and added msToLocalArrivalDepartureTimeString() just for the src/components/ArrivalDeparture.svelte page. - fixed lint issues - third round of PR feedback addressed
