+++ eslint.config.js
@@ -17,6 +17,7 @@ export default [
 				L: 'readonly',
 				google: 'readonly',
 				$state: 'readonly',
+				Temporal: 'readonly',
 				__SHOW_REGION_NAME_IN_NAV_BAR__: 'readonly',
 				__OBA_LOGO_URL_DARK__: 'readonly'
 			}
