+++ src/locales/ru.json
@@ -7,6 +7,10 @@
 		"miles": "миль"
 	},
 	"trip-planner": {
+		"arrive": "Прибыть",
+		"depart": "уйти",
+		"today": "Сегодня",
+		"tomorrow": "Завтра",
 		"plan_your_trip": "Планирование поездки",
 		"itinerary": "Маршрут",
 		"duration": "Продолжительность",
